% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_printing_functions.R
\name{create_CV_object}
\alias{create_CV_object}
\title{Create a CV_Printer object.}
\usage{
create_CV_object(
  data_location,
  pdf_mode = TRUE,
  sheet_is_publicly_readable = TRUE
)
}
\arguments{
\item{data_location}{Path of the spreadsheets holding all your data. This can be
either a URL to a Google Sheet with multiple sheets containing the four
data types or a path to a folder containing four \code{.csv}s with the neccesary
data.}

\item{pdf_mode}{Is the output being rendered into a pdf? Aka do links need
to be stripped?}

\item{sheet_is_publicly_readable}{If you're using google sheets for data,
is the sheet publicly available? (Makes authorization easier.)}

\item{source_location}{Where is the code to build your CV hosted?}
}
\value{
A new \code{CV_Printer} object.
}
\description{
Create a CV_Printer object.
}
