% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_printing_functions.R
\name{print_section}
\alias{print_section}
\title{Create a markdown code with given content and layout}
\usage{
print_section(
  cv,
  section_id,
  regex_expression = "{title} {description_bullets} \\\\hfill {start}"
)
}
\arguments{
\item{section_id}{ID of the entries section to be printed as encoded by the \code{section} column of the \code{entries} table}

\item{regex_expression}{Expression defining the output based on input from google drive og csv files. For example, use \code{{title}} to insert title. Use \code{print_var_names} to get a list of availible variable names.}
}
\value{
Markdown code with data from defined input.
}
\description{
Take a position data frame and the section id desired and prints the section to markdown.
}
