% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_printing_functions.R
\name{print_text_block}
\alias{print_text_block}
\title{Print text block}
\usage{
print_text_block(cv, label)
}
\arguments{
\item{label}{ID of the text block to print as encoded in \code{label} column of \code{text_blocks} table.}
}
\description{
Prints out text block identified by a given label.
}
