% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_printing_functions.R
\name{print_var_names}
\alias{print_var_names}
\title{Get list of variable names}
\usage{
print_var_names(cv, list_name)
}
\arguments{
\item{cv}{Enter name of dataframe where cv data is loaded. \code{CV} is default in the markdown template.}

\item{list_name}{List name (name of tab in Google Sheet). Default is \code{pdf_mode}, \code{links}, \code{entries_data}, \code{skills}, \code{text_block} and \code{contact_info}.}
}
\description{
Print available variable names.
}
