% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_csv_data_storage.R
\name{use_csv_data_storage}
\alias{use_csv_data_storage}
\title{Use CSVs for storing data}
\usage{
use_csv_data_storage(folder_name = "data")
}
\arguments{
\item{folder_name}{Name of the folder you want csvs stored in}
}
\value{
A new folder \verb{<folder_name>/} with \code{entries.csv}, \code{text_blocks.csv}, \code{language_skills.csv}, and \code{contact_info.csv} in it.
working directory.
}
\description{
Sets up examples of the four CSVs needed for building CV
}
\examples{

\dontrun{
  use_csv_data_storage(folder_name = "my_data")
}

}
