% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ddcv_template.R
\name{use_ddcv_template}
\alias{use_ddcv_template}
\title{Use template file from package}
\usage{
use_ddcv_template(
  file_name,
  params = NULL,
  output_file_name = file_name,
  output_dir = getwd(),
  create_output_dir = FALSE,
  warn_about_no_change = TRUE,
  open_after_making = FALSE
)
}
\arguments{
\item{file_name}{Name of file from templates to use: e.g. \code{cv.rmd}.}

\item{params}{Parameters used to fill in \code{whisker} template}

\item{output_file_name}{Name of file after being placed.}

\item{output_dir}{Directory location for output to be placed in.}

\item{create_output_dir}{If the requested output directory is missing should it be created?}

\item{warn_about_no_change}{If there is no change between the new file and what was already there, should a warning be issued?}

\item{open_after_making}{Should the file be opened after it has been written?}
}
\description{
Use template file from package
}
