% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Printer.R
\name{CV_Printer}
\alias{CV_Printer}
\title{R6 Class to print components of CV from data}
\description{
R6 Class to print components of CV from data

R6 Class to print components of CV from data
}
\details{
This class is initiated at the head of your CV or Resume Rmarkdown file and
then through various \verb{print_*} methods, builds the various components.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position_data}}{dataframe of positions by row}

\item{\code{position_data}}{dataframe of positions by row}

\item{\code{position_entry_template}}{\code{glue} template for building position entries}

\item{\code{pdf_mode}}{Is the output being rendered into a pdf? Aka do links need to be stripped?}

\item{\code{html_location}}{Where will the html version of your CV be hosted?}

\item{\code{pdf_location}}{Where will the pdf version of your CV be hosted?}

\item{\code{links}}{Internal array holding all the links that have been stripped in the order they were stripped.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CV_Printer$new()}}
\item \href{#method-set_pdf_mode}{\code{CV_Printer$set_pdf_mode()}}
\item \href{#method-sanitize_links}{\code{CV_Printer$sanitize_links()}}
\item \href{#method-strip_links_from_cols}{\code{CV_Printer$strip_links_from_cols()}}
\item \href{#method-print_section}{\code{CV_Printer$print_section()}}
\item \href{#method-print_text_block}{\code{CV_Printer$print_text_block()}}
\item \href{#method-print_skill_bars}{\code{CV_Printer$print_skill_bars()}}
\item \href{#method-print_links}{\code{CV_Printer$print_links()}}
\item \href{#method-print_contact_info}{\code{CV_Printer$print_contact_info()}}
\item \href{#method-print_link_to_other_format}{\code{CV_Printer$print_link_to_other_format()}}
\item \href{#method-set_style}{\code{CV_Printer$set_style()}}
\item \href{#method-clone}{\code{CV_Printer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a CV_Printer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$new(
  data_location,
  pdf_mode = FALSE,
  html_location,
  pdf_location,
  position_entry_template = default_position_entry_template,
  sheet_is_publicly_readable = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_location}}{Path of the spreadsheets holding all your data. This can be
either a URL to a google sheet with multiple sheets containing the four
data types or a path to a folder containing four \code{.csv}s with the neccesary
data.}

\item{\code{pdf_mode}}{Is the output being rendered into a pdf? Aka do links need to be stripped?}

\item{\code{html_location}}{What location will the HTML version of this CV be hosted at?}

\item{\code{pdf_location}}{What location will the PDF of this CV be hosted at?}

\item{\code{position_entry_template}}{A \code{glue} template for building position entries.}

\item{\code{sheet_is_publicly_readable}}{If you're using google sheets for data, is the sheet publicly available? (Makes authorization easier.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CV_Printer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_pdf_mode"></a>}}
\subsection{Method \code{set_pdf_mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$set_pdf_mode(pdf_mode = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sanitize_links"></a>}}
\subsection{Method \code{sanitize_links()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$sanitize_links(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-strip_links_from_cols"></a>}}
\subsection{Method \code{strip_links_from_cols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$strip_links_from_cols(data, cols_to_strip)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_section"></a>}}
\subsection{Method \code{print_section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$print_section(section_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_text_block"></a>}}
\subsection{Method \code{print_text_block()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$print_text_block(label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_skill_bars"></a>}}
\subsection{Method \code{print_skill_bars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$print_skill_bars(out_of = 5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_links"></a>}}
\subsection{Method \code{print_links()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$print_links()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_contact_info"></a>}}
\subsection{Method \code{print_contact_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$print_contact_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_link_to_other_format"></a>}}
\subsection{Method \code{print_link_to_other_format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$print_link_to_other_format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_style"></a>}}
\subsection{Method \code{set_style()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$set_style()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV_Printer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
