% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_csv_data_storage.R
\name{use_csv_data_storage}
\alias{use_csv_data_storage}
\title{Use CSVs for storing data}
\usage{
use_csv_data_storage(folder_name = "data", create_output_dir = TRUE)
}
\arguments{
\item{folder_name}{Name of the folder you want csvs stored in relative to current working directory}

\item{create_output_dir}{If the requested output directory is missing should it be created?}
}
\value{
A new folder \verb{<folder_name>/} with \code{entries.csv}, \code{text_blocks.csv}, \code{language_skills.csv}, and \code{contact_info.csv} in it.
working directory.
}
\description{
Sets up examples of the four CSVs needed for building CV
}
\examples{

# Make a temp directory for placing files
# This would be a real location for a typical situation
temp_dir <- fs::dir_create(fs::path(tempdir(), "cv_w_csvs"))

datadrivencv::use_csv_data_storage(
  folder_name = fs::path(temp_dir, "csv_data"),
  create_output_dir = TRUE
)

list.files(fs::path(temp_dir, "csv_data"))

}
