% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_datadriven_cv.R
\name{use_datadriven_cv}
\alias{use_datadriven_cv}
\title{Use Data Driven CV template}
\usage{
use_datadriven_cv(
  full_name = "Sarah Arcos",
  data_location = system.file("sample_data/", package = "datadrivencv"),
  pdf_location = "https://github.com/nstrayer/cv/raw/master/strayer_cv.pdf",
  html_location = "nickstrayer.me/datadrivencv/",
  source_location = "https://github.com/nstrayer/datadrivencv",
  use_network_logo = TRUE,
  open_files = TRUE
)
}
\arguments{
\item{full_name}{Your full name, used in title of document and header}

\item{data_location}{Path of the spreadsheets holding all your data. This can
be either a URL to a google sheet with multiple sheets containing the four
data types or a path to a folder containing four \code{.csv}s with the neccesary
data.}

\item{pdf_location}{What location will the PDF of this CV be hosted at?}

\item{html_location}{What location will the HTML version of this CV be hosted
at?}

\item{source_location}{Where is the code to build your CV hosted?}

\item{use_network_logo}{Should logo be an interactive network based on your
CV data? Note that this uses the function
\code{\link{build_network_logo()}} so will introduce a dependency on this
package.}

\item{open_files}{Should the added files be opened after creation?}
}
\value{
\code{cv.Rmd}, \code{dd_cv.css}, and \code{render_cv.R} written to the current
working directory.
}
\description{
Sets up the \code{.Rmd} file for a data-driven cv in current working directory.
Also adds css file for current CV so style can be custommized.
}
\examples{

\dontrun{
  use_datadriven_cv(
    full_name = "Nick Strayer",
    data_location = "https://docs.google.com/spreadsheets/d/14MQICF2F8-vf8CKPF1m4lyGKO6_thG-4aSwat1e2TWc",
    pdf_location = "https://github.com/nstrayer/cv/raw/master/strayer_cv.pdf",
    html_location = "nickstrayer.me/cv/",
    source_location = "https://github.com/nstrayer/cv"
  )
}

}
